backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
mistralai
#sentence-transformers
langchain
langchain-mistralai
numpy
httpx
fastmcp
